Revamping Notebook
- Its now possible to add section for a character from the list of
existing characters.
- New characters can be added to the notebook easily.
- Added ability to change character color and also remove a character
from the notebook.
- Added attachments panel for Character notes page
- Act and episode names are updated in notebook tree, if they are
changed by the user.
- Double clicking on a act, episode or scene node in the notebook tree
notonly expands that branch in the tree, but also sets the corresponding
break or scene current in the screenplay.

+++ src/document/attachments.cpp
@@ -60,7 +60,8 @@ Attachment::Attachment(QObject *parent)
 
 Attachment::~Attachment()
 {
-    QFile::remove(m_anonFilePath);
+    if(!m_anonFilePath.isEmpty() && QFile::exists(m_anonFilePath))
+        QFile::remove(m_anonFilePath);
     m_anonFilePath = QString();
 
     if(m_removeFileOnDelete)
@@ -237,6 +238,7 @@ Attachments::Attachments(QObject *parent)
 {
     connect(this, &Attachments::rowsMoved, this, &Attachments::attachmentsModified);
     connect(this, &Attachments::dataChanged, this, &Attachments::attachmentsModified);
+    connect(this, &Attachments::objectCountChanged, this, &Attachments::attachmentsModified);
     connect(this, &Attachments::objectCountChanged, this, &Attachments::attachmentCountChanged);
 }
